UT: Replace EventQueue with PyPubSubEventQueue
